ablations

+++ scripts/ablations/stepwise_ablation.py
@@ -35,7 +35,7 @@ if __name__ == "__main__":
 
     # python scripts/ablations/stepwise_ablation.py --output_key time_adaptation --number_training_runs 1 --data_start_percentage 10 --data_stop_percentage 100 --data_percentage_increment 10 --scenes forest,hilly,grassland --store_model_every_n_steps 100
 
-    # python scripts/ablations/stepwise_ablation.py --output_key learning_curve --number_training_runs 10 --data_start_percentage 100 --data_stop_percentage 100 --data_percentage_increment 10 --scenes forest --store_model_every_n_steps 100
+    # python scripts/ablations/stepwise_ablation.py --output_key learning_curve --number_training_runs 5 --data_start_percentage 100 --data_stop_percentage 100 --data_percentage_increment 10 --scenes forest --store_model_every_n_steps 100
 
     args = parser.parse_args()
 
@@ -72,39 +72,42 @@ if __name__ == "__main__":
 
     # Currently the model weights are stored every n steps.
     # This allows to reload the model and test it on the test dataloader.
+    train_and_delete = True
+    if train_and_delete:
+        shutil.rmtree(exp.general.model_path, ignore_errors=True)
 
-    shutil.rmtree(exp.general.model_path, ignore_errors=True)
+        Path(exp.general.model_path).mkdir(parents=True, exist_ok=True)
+        percent = range(
+            data_start_percentage, data_stop_percentage + data_percentage_increment, data_percentage_increment
+        )
 
-    Path(exp.general.model_path).mkdir(parents=True, exist_ok=True)
-    percent = range(data_start_percentage, data_stop_percentage + data_percentage_increment, data_percentage_increment)
+        with open(os.path.join(exp.general.model_path, "experiment_params.pkl"), "wb") as handle:
+            pickle.dump(exp, handle, protocol=pickle.HIGHEST_PROTOCOL)
 
-    with open(os.path.join(exp.general.model_path, "experiment_params.pkl"), "wb") as handle:
-        pickle.dump(exp, handle, protocol=pickle.HIGHEST_PROTOCOL)
+        # Train model in various configurations and the validation results per epoch are returned in results_epoch.
+        results_epoch = {}
+        for scene in scenes:
+            exp.ablation_data_module.env = scene
+            percentage_results = {}
+            for percentage in percent:
+                exp.ablation_data_module.training_data_percentage = percentage
+                run_results = {}
+                for run in range(number_training_runs):
+                    exp.general.store_model_every_n_steps_key = f"ablation_{output_key}_{scene}_{percentage}_{run}"
+                    res, _ = training_routine(exp, seed=run)
+                    run_results[f"run_{run}"] = copy.deepcopy(res)
+                    torch.cuda.empty_cache()
+                percentage_results[f"percentage_{percentage}"] = copy.deepcopy(run_results)
+            results_epoch[scene] = copy.deepcopy(percentage_results)
 
-    # Train model in various configurations and the validation results per epoch are returned in results_epoch.
-    results_epoch = {}
-    for scene in scenes:
-        exp.ablation_data_module.env = scene
-        percentage_results = {}
-        for percentage in percent:
-            exp.ablation_data_module.training_data_percentage = percentage
-            run_results = {}
-            for run in range(number_training_runs):
-                exp.general.store_model_every_n_steps_key = f"ablation_{output_key}_{scene}_{percentage}_{run}"
-                res, _ = training_routine(exp, seed=run)
-                run_results[f"run_{run}"] = copy.deepcopy(res)
-                torch.cuda.empty_cache()
-            percentage_results[f"percentage_{percentage}"] = copy.deepcopy(run_results)
-        results_epoch[scene] = copy.deepcopy(percentage_results)
-
-    # Store epoch output to disk.
-    p = os.path.join(exp.general.model_path, f"{output_key}_epochs.pkl")
-    try:
-        os.remove(p)
-    except OSError as error:
-        pass
-    with open(p, "wb") as handle:
-        pickle.dump(results_epoch, handle, protocol=pickle.HIGHEST_PROTOCOL)
+        # Store epoch output to disk.
+        p = os.path.join(exp.general.model_path, f"{output_key}_epochs.pkl")
+        try:
+            os.remove(p)
+        except OSError as error:
+            pass
+        with open(p, "wb") as handle:
+            pickle.dump(results_epoch, handle, protocol=pickle.HIGHEST_PROTOCOL)
 
     # Test all stored models on the test dataloader and store the results.
     exp.general.skip_train = True
